requirements.txt
openpyxl>=3.1.0
pandas>=1.5.0
PySide6>=6.6.0
Pillow
cairosvg 